chore: moved DocsMod to ftb lib & a bit of side-safety
