--- common/src/main/java/dev/ftb/mods/ftbquests/integration/DocsMod.java
@@ -1,14 +1,0 @@
-package dev.ftb.mods.ftbquests.integration;
-
-public interface DocsMod {
-    void openDocsPage(String path);
-
-    enum None implements DocsMod {
-        INSTANCE;
-
-        @Override
-        public void openDocsPage(String path) {
-
-        }
-    }
-}
